Move beluga neighbor stuff into its own file
